server: Refactor socket server setup
Change-type: patch
