+++ lib/server/web-server.js
@@ -20,25 +20,19 @@ const logger = require('../logger').getLogger(__filename)
 const environment = require('../environment')
 const express = require('express')
 const http = require('http')
-const socketIo = require('socket.io')
 
 const middlewares = require('./http/middlewares')
 const routes = require('./http/routes')
-const {
-	bindRoutes
-} = require('./routes')
+const socket = require('./socket')
 
 exports.createWebServer = (context, jellyfish, queue, configuration) => {
 	const application = express()
 	const server = http.Server(application)
-	const socketServer = socketIo(server)
+	const socketServer = socket(jellyfish, server)
 
 	application.set('port', configuration.http.port)
 	middlewares(application, jellyfish)
 	routes(application, jellyfish, queue)
-
-	// Socket server routes
-	bindRoutes(jellyfish, application, socketServer)
 
 	// Resolve the promise once the server is ready to accept incoming
 	// connections.
